+++ lib/features/history/presentation/history_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:math_skill_up/core/components/app_bar.dart';
+import 'package:math_skill_up/core/components/common_scaffold.dart';
 import 'package:math_skill_up/features/history/presentation/history_chart.dart';
 import 'package:math_skill_up/features/history/presentation/history_list.dart';
 import 'package:math_skill_up/features/history/presentation/history_setting_bar.dart';
@@ -11,21 +11,16 @@ class HistoryScreen extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return Scaffold(
-      appBar: CommonAppBar(titleText: '히스토리'),
-      body: Padding(
-        padding: const EdgeInsets.only(
-            top: 8.0, bottom: 16.0, left: 16.0, right: 16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            HistorySettingBar(),
-            const SizedBox(height: 16),
-            OptionalHistoryChart(),
-            const SizedBox(height: 16),
-            Expanded(child: HistoryList()),
-          ],
-        ),
+    return CommonScaffold(
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          HistorySettingBar(),
+          const SizedBox(height: 16),
+          OptionalHistoryChart(),
+          const SizedBox(height: 16),
+          Expanded(child: HistoryList()),
+        ],
       ),
     );
   }
