+++ lib/features/result/presentation/result_screen.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:math_skill_up/core/components/app_bar.dart';
+import 'package:math_skill_up/core/components/common_scaffold.dart';
 
 class ResultScreen extends StatelessWidget {
   const ResultScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: CommonAppBar(titleText: '결과'),
-    );
+    return CommonScaffold(child: Text("temp"));
   }
 }
